refactor(mcp-kit): eliminate format-tool-result.ts — logic and type owned by interceptor
Move formatToolResult logic and ToolWireResult type into McpSerializationInterceptor,
which becomes the single owner. McpExceptionFilter imports ToolWireResult from
interceptors/. Remove formatToolResult from the public API surface.
