+++ packages/mcp-kit/src/serialization/index.ts
@@ -4,5 +4,3 @@ export { McpToolResult } from './mcp-tool-result';
 export { McpResourceResult } from './mcp-resource-result';
 export type { ResourceContent } from './mcp-resource-result';
 export { ToolError, ResourceError, PromptError } from './tool-errors';
-export { formatToolResult } from './format-tool-result';
-export type { ToolWireResult } from './format-tool-result';
